refactor: Extract LiquidfunRenderer
